+++ src/pages/[id]/editteam/index.tsx
@@ -26,7 +26,10 @@ import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
 const teamSchema: z.ZodSchema<PostGroupRequest> = z.object({
-  name: z.string().min(1, { message: '팀 이름은 필수 입력입니다.' }),
+  name: z
+    .string()
+    .min(1, { message: '팀 이름은 필수 입력입니다.' })
+    .max(30, { message: '팀 이름은 30글자 이하로 입력해주세요.' }),
   image: z.string().min(1),
 });
