+++ src/pages/[id]/index.tsx
@@ -9,7 +9,6 @@ import Button, {
   TextSize,
 } from '@/components/common/Button/Button';
 import DropDown from '@/components/common/Dropdown';
-import Input from '@/components/common/Input';
 import { Modal } from '@/components/modal';
 import TaskLists from '@/components/TaskList/TaskLists';
 import Members from '@/components/Team/Members';
@@ -18,27 +17,22 @@ import { useToast } from '@/hooks/useToast';
 import {
   useDeleteTeamMutation,
   useInviteGroupQuery,
-  useTaskListMutation,
   useTasksQuery,
   useTeamQuery,
 } from '@/queries/groups.queries';
-
 import Image from 'next/image';
 import { useRouter } from 'next/router';
-import { useRef } from 'react';
 import WithOutTeam from '../withoutteam';
 
 export default function TeamPage() {
   const router = useRouter();
   const today = new Date().toISOString().split('T')[0];
   const { id } = router.query;
-  const { data: group, isError } = useTeamQuery(Number(id));
+  const { data: group, isError, isFetched } = useTeamQuery(Number(id));
   const { data: inviteLink } = useInviteGroupQuery(Number(id));
   const { data: tasks } = useTasksQuery({ id: Number(id), date: today });
 
-  const createTaskList = useTaskListMutation();
   const deleteTeam = useDeleteTeamMutation();
-  const taskListNameRef = useRef<HTMLInputElement>(null);
   const { toast } = useToast();
 
   const TODAY_PROGRESS = tasks?.length
@@ -46,6 +40,13 @@ export default function TeamPage() {
     : 0;
   const TODAY_PROGRESS_PERCENT = Math.floor(TODAY_PROGRESS * 100);
 
+  if (!isFetched) {
+    return (
+      <div className="flex h-[50rem] items-center justify-center">
+        <p className="text-4xl">로딩 중 입니다....</p>
+      </div>
+    );
+  }
   if (isError || !group) {
     return <WithOutTeam />;
   }
@@ -96,39 +97,6 @@ export default function TeamPage() {
     });
   };
 
-  const handleCreateTask = () => {
-    const taskListName = taskListNameRef.current?.value.trim() || '';
-
-    if (taskListName) {
-      createTaskList.mutate(
-        { groupId: Number(id), name: taskListName },
-        {
-          onSuccess: () => {
-            toast({
-              title: '목록생성 완료',
-              description: '새 목록이 생성되었습니다',
-            });
-          },
-          onError: () => {
-            toast({
-              title: '목록생성 실패',
-              variant: 'destructive',
-            });
-          },
-        }
-      );
-      if (taskListNameRef.current) {
-        taskListNameRef.current.value = '';
-      }
-    } else {
-      toast({
-        title: '목록생성 실패',
-        description: '할 일 목록명을 입력해주세요.',
-        variant: 'destructive',
-      });
-    }
-  };
-
   return (
     <div className="flex w-full flex-col gap-5 px-20 pt-10 tab:px-5">
       <div
@@ -176,50 +144,6 @@ export default function TeamPage() {
           </div>
         </div>
       </div>
-      <div className="flex justify-between">
-        <div className="flex gap-2">
-          <p className="text-lg-medium">할 일 목록</p>
-          <p className="text-lg-medium text-default">
-            ({group.taskLists.length}개)
-          </p>
-        </div>
-
-        <Modal>
-          <Modal.Toggle className="text-brand-primary">
-            + 새로운 목록 추가하기
-          </Modal.Toggle>
-          <Modal.Portal>
-            <Modal.Overlay />
-            <Modal.Content withToggle>
-              <div className="flex flex-col gap-5">
-                <Modal.Header>
-                  <Modal.Title>할 일 목록</Modal.Title>
-                </Modal.Header>
-                <Input
-                  id="task-list-name"
-                  wrapperClassName="w-[280px]"
-                  placeholder="목록 명을 입력해주세요"
-                  ref={taskListNameRef}
-                />
-                <Modal.Close>
-                  <Button
-                    buttonStyle={ButtonStyle.Box}
-                    textColor={TextColor.White}
-                    textSize={TextSize.Large}
-                    buttonWidth={ButtonWidth.Full}
-                    buttonBackgroundColor={ButtonBackgroundColor.Green}
-                    buttonBorderColor={ButtonBorderColor.Green}
-                    buttonPadding={ButtonPadding.Medium}
-                    onClick={handleCreateTask}
-                  >
-                    만들기
-                  </Button>
-                </Modal.Close>
-              </div>
-            </Modal.Content>
-          </Modal.Portal>
-        </Modal>
-      </div>
       <TaskLists taskLists={group.taskLists} id={id!.toString()} />
       <p className="text-lg-medium">리포트</p>
       <div
@@ -227,7 +151,6 @@ export default function TeamPage() {
      justify-between rounded-xl bg-secondary px-5 mob:gap-5"
       >
         <CircularProgressChart value={TODAY_PROGRESS_PERCENT} />
-
         <div className="flex w-[25rem] flex-col gap-5 tab:w-[17.5rem]">
           <div
             className="flex h-[4.78125rem]
